move integration tests to e2e package
