+++ e2e/testdata/modcmd/nested.tar.gz
