+++ e2e/testdata/modcmd/vendored.tar.gz
