+++ e2e/testdata/snapshots/TestModCmdNested
