+++ e2e/testdata/snapshots/TestModCmdNoDependencies
