+++ e2e/testdata/snapshots/TestModCmdVendored
